--- internal/server/application/README.md
@@ -1,10 +1,0 @@
-# internal/service
-
-Пакет `service` содержит бизнес-логику и играет ключевую роль в реализации функциональности приложения.
-
-В нём описаны правила, процессы и операции, которые определяют поведение приложения.
-
-Принципы организации:
-- Сервисы должны быть независимы от деталей транспорта (HTTP, gRPC и т.д.).
-- Взаимодействие с базой данных происходит через интерфейсы репозиториев.
-- Каждый сервис должен иметь четко определенную область ответственности.
