--- internal/server/domain/metric_repository.go
@@ -1,6 +1,0 @@
-package domain
-
-type MetricRepository interface {
-	Save(metrics *Metrics)
-	GetMetric(metricType, metricName string) *Metrics
-}
