+++ internal/server/domain/model/metrics.go
@@ -1,4 +1,4 @@
-package domain
+package model
 
 import (
 	"errors"
