+++ internal/server/domain/repository/metric_repository.go
@@ -1,0 +1,10 @@
+package repository
+
+import (
+	"github.com/sirajDeveloper/metrics-alerts-collector/internal/server/domain"
+)
+
+type MetricRepository interface {
+	Save(metrics *domain.Metrics)
+	GetMetric(metricType, metricName string) *domain.Metrics
+}
