+++ internal/server/handler/http/routes.go
@@ -1,4 +1,4 @@
-package handler
+package http
 
 import (
 	"log"
